// packages/base/lib/interfaces/api.ts
export interface APIPreferences extends ApiItem {
  recordingExpireDays: number;
  type: 'customerpreferences';
}
// ...
export interface APICallBundle extends ApiItem {
  autoRenew: boolean;
  availableMins: number;
  includedMins: number;
  monthlyMins: number;
  name: string;
  priceMonthly: number;
  status: 'ACTIVE' | 'EXPIRED';
  type: 'customerbundle';
  usedMins: number;
  validFrom: Date;
  validTo: Date;
}

export interface APIAvailableBundle extends ApiItem {
  destinations: string[];
  inclusiveMins: number;
  name: string;
  priceMonthly: number;
  type: 'availablebundle';
}

export interface APICallRecording extends ApiItem {
  direction: 'IN' | 'OUT';
  callId: string;
  linkedId: string;
  endpoint: string;
  partyId: string;
  size: number;
  started: string;
  type: 'recording';
}

export interface APICall extends ApiItem {
  type: 'call';
  scope: 'LOCAL' | 'DOMAIN' | 'EXTERNAL';
  direction: 'IN' | 'OUT';
  from: string;
  to: string;
  callerId: string;
  callStarted: string;
  outcome: 'ANSWERED' | 'NO_ANSWER' | 'BUSY' | 'FAILED';
  duration: number;
  cost: number;
  sell: number;
  callId: string;
  linkedId: string;
  srcEndpoint: string;
  dstEndpoint: string;
}

export enum AccountType {
  BUSINESS = 'BUSINESS',
  RESIDENTIAL = 'RESIDENTIAL',
  UNKNOWN = 'UNKNOWN',
}

export interface APICustomer extends ApiItem {
  type: 'customer';
  accountType: AccountType;
  id: string;
  created: Date;
  company: string;
  firstName: string;
  lastName: string;
  telephone: string;
  email: string;
  address1: string;
  address2: string;
  city: string;
  postcode: string;
  enabled: boolean;
  partnerId?: string;
  partnerCompany?: string;
  userEmailUpdatable: boolean;
}
// ...
export interface APIBaseLinkedUser {
  // common properties for linked user (SUPPORT,RESELLER,ADMIN)
  type: 'linkeduser';
  accessLevel:
    | 'CUSTOMER'
    | 'EXTENSION'
    | 'RESELLER'
    | 'SUPPORT_L1'
    | 'SUPPORT_L2'
    | 'ADMIN';
  email: string;
  enabled: boolean;
  pending: boolean;
  owner: boolean;
  defaultExtension?: string;
  recordingAccess: 'NONE' | 'OWN' | 'ALL';
}

export interface APICreateLinkedUser extends APIBaseLinkedUser {
  activateUrl: string;
}

export interface APILinkedUser extends APIBaseLinkedUser, ApiItem {
  type: 'linkeduser';
  userId?: string;
  firstName?: string;
  lastName?: string;
  username?: string;

  // properties for CUSTOMER, EXTENSION
  acceptedInvite?: Date;
}

export interface APIOutgoingCLI extends ApiItem {
  allowCalls: boolean;
  allowSms: boolean;
  number: string;
  status: 'APPROVED' | 'PENDING' | 'REJECTED';
  type: 'outgoingcallerid';
  uri: string;
}

export interface APIPhonebook extends ApiItem {
  email: string;
  name: string;
  phoneNumber: string;
  priority: number;
  speedDial: number;
  type: 'phonebookentry';
}
// ...
export interface APISMSMessage extends ApiItem {
  body: string;
  cost: number;
  creditsUsed: number;
  direction: 'IN' | 'OUT';
  from: string;
  to: string;
  type: 'smsmessage';
  sendStatus?: 'PENDING' | 'SENT' | 'FAILED' | 'PARTIAL';
}

export interface APISoundPrompt extends ApiItem {
  type: 'prompt';
  name: string;
  size: number;
}

export interface APISoundMusic extends ApiItem {
  type: 'music';
  name: string;
  size: number;
  enabled: boolean;
}

interface APIBaseTimeInterval extends ApiItem {
  type: 'timeinterval';
  name: string;
  shared: boolean;
}

export interface WeeklyIncludedTime {
  daysOfWeek: DayOfWeek[];
  startTime: string;
  endTime: string;
}

export interface YearlyIncludedTime {
  month: MonthAbbreviation;
  dayOfMonth: number;
  endTime: string;
  startTime: string;
}

export interface APIWeeklyTimeInterval extends APIBaseTimeInterval {
  repeat: 'WEEKLY';
  includedTimes: WeeklyIncludedTime[];
}

export interface APIYearlyTimeInterval extends APIBaseTimeInterval {
  repeat: 'YEARLY';
  includedTimes: YearlyIncludedTime[];
}

export type APITimeInterval = APIWeeklyTimeInterval | APIYearlyTimeInterval;